ovn-docker-overlay-driver: Get upto-date with Docker's libnetwork.
Docker has finalized its plugin model. It is supposed to be a RPC
model in the sense that the datastructures used inside
libnetwork will be converted to json and sent over http. So it is easier
to write the plugin in golang by using the same packages as libnetwork uses.
But in OVN, we use OpenStack Neutron APIs and the client for it is Python.

So the options are:
1. Write a small golang library to create Neutron REST API calls.
2. Hope that Docker's libnetwork won't change its datastructures
once the API stabilizes in a backwards incompatible way and write
a Python webserver to handle libnetwork's API calls.

The second option is much simpler to begin with. So this commit
continues on that model.

The API nevertheless will change soon to handle Docker labels. So this
is not yet ready to be upstreamed to OVS repo.

Signed-off-by: Gurucharan Shetty <shettyg@nicira.com>